* 'hotfix/fixbug' of github.com:zhiyicx/thinksns-plus-android:
  (#1325) code fix this

* 'hotfix/fixbug' of github.com:zhiyicx/thinksns-plus-android:
  (#1325) code fix this

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/channel/group_dynamic/GroupDynamicDetailFragment.java
@@ -332,7 +332,7 @@ public class GroupDynamicDetailFragment extends TSListFragment<GroupDynamicDetai
     }
 
     @Override
-    public void setSpanText(int position, int note, int amount, TextView view, boolean canNotRead) {
+    public void setSpanText(int position, int note, long amount, TextView view, boolean canNotRead) {
         initImageCenterPopWindow(position, (float) amount,
                 note, R.string.buy_pay_words_desc, false);
     }
