+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/q_a/detail/question/QuestionDetailFragment.java
@@ -166,7 +166,7 @@ public class QuestionDetailFragment extends TSListFragment<QuestionDetailContrac
     }
 
     @Override
-    public void setSpanText(int position, int note, int amount, TextView view, boolean canNotRead) {
+    public void setSpanText(int position, int note, long amount, TextView view, boolean canNotRead) {
         onToWatchClick(null, position, canNotRead);
     }
